Extract Shield enforcement collaborator from GovernancePipeline
Move the runtime-enforcement surface (preflight/postflight gate chains,
allow/deny recording, gate-context construction, and the single gate-channel
counter writer _observe_completed_call) out of GovernancePipeline into a new
Shield collaborator. The pipeline now composes Shield via constructor
injection and forwards _run_preflight/_run_postflight/_score_and_gate_preflight/
_enforce_postflight/_apply_postflight_to_output to it, keeping the public gate_*
adapters and external callers (gate/server, tests) unchanged.

Shield owns its own lock, reads live policy via an injected provider, and scores
via an injected scorer (DIP). No behavior change: 1884 passed / 4 skipped.
