+++ src/tracemill/governance/pipeline.py
@@ -10,6 +10,7 @@ from typing import TYPE_CHECKING, Literal
 
 from tracemill.governance.assessor import DefaultAssessor
 from tracemill.governance.registry import SessionRegistry
+from tracemill.governance.shield import Shield
 from tracemill.governance.results import (
     Evidence,
     EvidencePointer,
@@ -34,12 +35,7 @@ if TYPE_CHECKING:
         ToolCallEvent,
     )
     from tracemill.sdk.gate_policy import GatePolicy
-    from tracemill.sdk.gate_types import (
-        GateContext,
-        PostflightVerdict,
-        ToolCallRequest,
-        ToolCallResult,
-    )
+    from tracemill.sdk.gate_types import PostflightVerdict
     from tracemill.sdk.verdict import Verdict
 
 
@@ -81,8 +77,6 @@ class GovernancePipeline:
         project_root: str | None = None,
         policy: "GatePolicy | None" = None,
     ) -> None:
-        import threading
-
         self._store = store
         self._labeler = labeler
         self._budget = budget_tracker
@@ -91,9 +85,9 @@ class GovernancePipeline:
         self._thresholds = thresholds
         self._project_root = project_root
         self.policy: "GatePolicy | None" = policy
-        self._gate_lock = threading.Lock()
         self._registry = SessionRegistry(store)
         self._assessor = DefaultAssessor(labeler, rules, engine)
+        self._shield = Shield(self._registry, lambda: self.policy, self._score_event)
         self._write_failures: dict[str, int] = {}  # session_id → consecutive failure count
         self._MAX_WRITE_FAILURES = 10
         self._phase23_attempts: dict[str, int] = {}  # source_event_key → attempt count
@@ -539,112 +533,9 @@ class GovernancePipeline:
 
     # ─── Central gate execution helpers ───────────────────────────────────────
 
-    def _build_gate_context(self, session_id: str) -> "GateContext":
-        """Build a GateContext from current session state.
-
-        Reads only — the counter and decision log are owned by SessionState and
-        exposed through methods. The shield never mutates state from here.
-        """
-        from tracemill.sdk.gate_types import GateContext
-
-        state = self._registry.get(session_id)
-        if state is None:
-            return GateContext(session_id=session_id, tool_call_count=0, denied_count=0)
-
-        return GateContext(
-            session_id=session_id,
-            tool_call_count=state.tool_call_count,
-            denied_count=state.denied_count,
-            prior_verdicts=state.prior_verdicts(),
-            prior_tool_call_ids=state.prior_tool_call_ids(),
-        )
-
-    def _to_tool_call_request(self, trace: "EventTrace") -> "ToolCallRequest":
-        """Convert a fully-assessed EventTrace into a gate-facing ToolCallRequest."""
-        from tracemill.sdk.gate_types import ToolCallRequest
-
-        return ToolCallRequest(
-            tool=trace.canonical_tool or trace.tool_name or "unknown",
-            input=trace.tool_input,
-            target=trace.target_resource,
-            mechanism=trace.mechanism or "unknown",
-            effect=trace.effect or "read_only",
-            capabilities=trace.capability,
-            scope=trace.scope,
-            role=trace.role,
-            action=trace.action,
-            risk_score=trace.risk_score or 0,
-            risk_band=trace.risk_band or "unknown",
-            suggested_action=trace.suggested_action or "allow",
-            reason=trace.reason or "",
-            session_id=trace.session_id,
-            tool_call_id=trace.tool_call_id,
-            event_trace=trace,
-        )
-
-    def _to_tool_call_result(
-        self,
-        trace: "EventTrace",
-        *,
-        output: dict | None = None,
-        duration_ms: int | None = None,
-        error: str | None = None,
-    ) -> "ToolCallResult":
-        """Convert a fully-assessed EventTrace + output into a gate-facing ToolCallResult."""
-        from tracemill.sdk.gate_types import ToolCallResult
-        from tracemill.trace import _deep_freeze, EMPTY_MAP
-
-        return ToolCallResult(
-            tool=trace.canonical_tool or trace.tool_name or "unknown",
-            input=trace.tool_input,
-            target=trace.target_resource,
-            output=_deep_freeze(output) if output else EMPTY_MAP,
-            duration_ms=duration_ms,
-            error=error,
-            mechanism=trace.mechanism or "unknown",
-            effect=trace.effect or "read_only",
-            capabilities=trace.capability,
-            risk_score=trace.risk_score or 0,
-            risk_band=trace.risk_band or "unknown",
-            suggested_action=trace.suggested_action or "allow",
-            reason=trace.reason or "",
-            session_id=trace.session_id,
-            tool_call_id=trace.tool_call_id,
-            event_trace=trace,
-        )
-
     def _run_preflight(self, trace: "EventTrace", *, session_id: str) -> "Verdict":
-        """Execute the preflight gate chain. Returns first DENY or ALLOW.
-
-        All gates come from the GatePolicy. No per-method overrides.
-        Fail-closed: if any gate or internal logic raises, returns DENY.
-        """
-        from tracemill.sdk.verdict import Verdict
-
-        try:
-            request = self._to_tool_call_request(trace)
-            ctx = self._build_gate_context(session_id)
-        except Exception as exc:
-            deny = Verdict.deny(f"gate setup error (fail-closed): {type(exc).__name__}: {exc}")
-            self._record_denial(session_id, deny)
-            return deny
-
-        # Run policy chain
-        if self.policy and self.policy.has_preflight:
-            for gate in self.policy.preflight_gates:
-                try:
-                    verdict = gate(request, ctx)
-                except Exception as exc:
-                    # Fail-closed: gate exception = DENY
-                    deny = Verdict.deny(f"gate error (fail-closed): {type(exc).__name__}: {exc}")
-                    self._record_denial(session_id, deny)
-                    return deny
-                if verdict.denied:
-                    self._record_denial(session_id, verdict)
-                    return verdict
-
-        self._record_allow(session_id, tool_call_id=trace.tool_call_id)
-        return Verdict.allow()
+        """Forward to the Shield's preflight enforcement chain."""
+        return self._shield.run_preflight(trace, session_id=session_id)
 
     def _run_postflight(
         self,
@@ -655,64 +546,14 @@ class GovernancePipeline:
         duration_ms: int | None = None,
         error: str | None = None,
     ) -> "PostflightVerdict":
-        """Execute the postflight gate chain. Returns most restrictive action.
-
-        Priority: TERMINATE > SUPPRESS > REDACT > ALERT > ACCEPT.
-        Fail-closed: if any gate or setup raises, returns SUPPRESS.
-        """
-        from tracemill.sdk.gate_types import PostflightAction, PostflightVerdict
-
-        try:
-            result = self._to_tool_call_result(
-                trace, output=output, duration_ms=duration_ms, error=error
-            )
-            ctx = self._build_gate_context(session_id)
-        except Exception as exc:
-            return PostflightVerdict(
-                action=PostflightAction.SUPPRESS,
-                reason=f"postflight setup error (fail-closed): {type(exc).__name__}: {exc}",
-            )
-
-        # Action severity ordering
-        _SEVERITY = {
-            PostflightAction.ACCEPT: 0,
-            PostflightAction.ALERT: 1,
-            PostflightAction.REDACT: 2,
-            PostflightAction.SUPPRESS: 3,
-            PostflightAction.TERMINATE: 4,
-        }
-
-        most_severe = PostflightVerdict()
-
-        # Run policy chain
-        if self.policy and self.policy.has_postflight:
-            for gate in self.policy.postflight_gates:
-                try:
-                    pv = gate(result, ctx)
-                except Exception as exc:
-                    # Fail-closed: gate exception = SUPPRESS (not TERMINATE to avoid crashing)
-                    pv = PostflightVerdict(
-                        action=PostflightAction.SUPPRESS,
-                        reason=f"postflight gate error (fail-closed): {type(exc).__name__}: {exc}",
-                    )
-                if _SEVERITY.get(pv.action, 0) > _SEVERITY.get(most_severe.action, 0):
-                    most_severe = pv
-
-        return most_severe
-
-    def _record_denial(self, session_id: str, verdict: "Verdict") -> None:
-        """Record a shield denial. Does not touch the tool-call counter."""
-        self._ensure_gate_state(session_id).record_denial(verdict)
-
-    def _record_allow(self, session_id: str, *, tool_call_id: str | None = None) -> None:
-        """Record a shield allow. The tool-call counter is NOT advanced here —
-        it advances only when the allowed call is observed at completion
-        (see _observe_completed_call). The shield only reads the counter."""
-        self._ensure_gate_state(session_id).record_allow(tool_call_id=tool_call_id)
-
-    def _ensure_gate_state(self, session_id: str):
-        """Return session state for gate context tracking (thread-safe, ephemeral)."""
-        return self._registry.ensure(session_id)
+        """Forward to the Shield's postflight enforcement chain."""
+        return self._shield.run_postflight(
+            trace,
+            session_id=session_id,
+            output=output,
+            duration_ms=duration_ms,
+            error=error,
+        )
 
     def score_tool_call_event(self, event: "tracemill.types.SessionEvent") -> "SessionMeta":
         """Score an enriched SessionEvent via the canonical bridge.
@@ -1622,17 +1463,8 @@ class GovernancePipeline:
     # Postflight verdicts (SUPPRESS/TERMINATE/REDACT) are enforced via framework-native signals.
 
     def _score_and_gate_preflight(self, payload: dict) -> tuple:
-        """Score a tool call and run the preflight gate chain. Thread-safe.
-
-        Returns (trace, verdict) tuple. Verdict is ALLOW or DENY.
-        Session ID is derived from the trace (which normalizes from payload).
-        """
-        with self._gate_lock:
-            trace = self._score_event(payload)
-        # Use trace.session_id for consistency — it normalizes empty/missing values
-        session_id = trace.session_id
-        verdict = self._run_preflight(trace, session_id=session_id)
-        return trace, verdict
+        """Forward to the Shield: score a tool call and run the preflight chain."""
+        return self._shield.score_and_gate_preflight(payload)
 
     def _enforce_postflight(
         self,
@@ -1643,19 +1475,8 @@ class GovernancePipeline:
         duration_ms: int | None = None,
         error: str | None = None,
     ) -> "PostflightVerdict":
-        """Observe the completed call, then run the postflight chain.
-
-        This is the post-execution point: the call the shield allowed has now
-        run, so the trace observes it here — the single place the tool-call
-        counter advances in the gate channel. Then callers check verdict.action:
-          - ACCEPT: return result normally
-          - REDACT: strip keys from output before returning
-          - SUPPRESS: return empty/neutral output to the model
-          - TERMINATE: raise/signal session termination
-          - ALERT: return result normally but log alert
-        """
-        self._observe_completed_call(trace, session_id=session_id)
-        return self._run_postflight(
+        """Forward to the Shield: observe the completed call, then run postflight."""
+        return self._shield.enforce_postflight(
             trace,
             session_id=session_id,
             output=output,
@@ -1663,44 +1484,10 @@ class GovernancePipeline:
             error=error,
         )
 
-    def _observe_completed_call(self, trace: "EventTrace", *, session_id: str) -> None:
-        """Advance the single tool-call counter for a shield-allowed call that
-        has now executed. The trace only ever observes what the gate allowed, so
-        this is the one writer of the counter in the gate channel."""
-        state = self._ensure_gate_state(session_id)
-        phase_window = state.snapshot().phase_window
-        state.increment_budget(
-            mechanism=trace.mechanism,
-            effect=trace.effect,
-            scope=frozenset(trace.scope),
-            role=frozenset(trace.role),
-            action=frozenset(trace.action),
-            capability=frozenset(trace.capability),
-            structure=frozenset(trace.structure),
-            phase=phase_window[-1] if phase_window else None,
-        )
-
     @staticmethod
     def _apply_postflight_to_output(pv: "PostflightVerdict", result: str) -> str:
-        """Apply a postflight verdict to a string result.
-
-        Returns the (possibly redacted/suppressed) output string.
-        Raises RuntimeError on TERMINATE.
-        """
-        from tracemill.sdk.gate_types import PostflightAction
-
-        if pv.action == PostflightAction.ACCEPT or pv.action == PostflightAction.ALERT:
-            return result
-        if pv.action == PostflightAction.SUPPRESS:
-            return "[output suppressed by policy]"
-        if pv.action == PostflightAction.REDACT:
-            redacted = result
-            for key in pv.redaction_keys:
-                redacted = redacted.replace(key, "[REDACTED]")
-            return redacted
-        if pv.action == PostflightAction.TERMINATE:
-            raise RuntimeError(f"Session terminated by policy: {pv.reason}")
-        return result
+        """Forward to the Shield's postflight-output transform."""
+        return Shield.apply_postflight_to_output(pv, result)
 
     def gate_crewai(self) -> None:
         """Register tracemill into CrewAI's before/after tool_call hooks.
